Better global styling. Got rid of mega index.js file and moved most styling back to the components
