+++ mishmash/frontend/src/components/TopNavBar.js
@@ -1,94 +1,140 @@
-/**
- * Study Abroad Program - Top Navigation Bar Component
- * =============================================
- * 
- * This component provides the main navigation header for the application.
- * It displays the application title, user information, and navigation controls
- * based on authentication state.
- * 
- * Features:
- * - Responsive navigation header
- * - Dynamic content based on auth state
- * - User welcome message
- * - Login/Logout functionality
- * - Dashboard navigation
- * 
- * States:
- * - Authenticated: Shows username, dashboard link, and logout
- * - Unauthenticated: Shows login button
- */
-
 import React from 'react';
+import { styled, useTheme } from '@mui/material/styles';
 import { useAuth } from '../context/AuthContext';
 import { useNavigate } from 'react-router-dom';
 import axiosInstance from '../utils/axios';
-import {
-    NavBar,
-    NavLogo,
-    NavLogoImage,
-    NavTitle,
-    NavControls,
-    NavButton,
-    WelcomeText,
-} from './styled';
 
-/**
- * Top Navigation Bar Component
- * 
- * @param {function} onLoginClick - Callback function for login button click
- * @returns {React.ReactElement} The navigation bar component
- */
+// -------------------- STYLES --------------------
+const NavBar = styled('div')(({ theme }) => ({
+  backgroundColor: theme.palette.primary.main,
+  padding: '1rem',
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  color: theme.palette.primary.contrastText,
+  boxShadow: theme.shadows.card,
+  position: 'fixed',
+  top: 0,
+  left: 0,
+  right: 0,
+  zIndex: 1000,
+}));
+
+const NavLogo = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  gap: '12px',
+  cursor: 'pointer',
+});
+
+const NavLogoImage = styled('img')({
+  height: '40px',
+  width: 'auto',
+});
+
+const NavTitle = styled('span')(({ theme }) => ({
+  fontSize: theme.typography.h4.fontSize,
+  fontWeight: theme.typography.h4.fontWeight,
+  textShadow: theme.textShadows.subtle,
+}));
+
+const NavControls = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  gap: '20px',
+});
+
+const NavButton = styled('button')(({ theme, variant = 'default' }) => {
+  const getStyles = () => {
+    switch (variant) {
+      // case 'transparent':
+      //   return {
+      //     backgroundColor: 'transparent',
+      //     border: `1px solid ${theme.palette.overlay.subtle}`,
+      //     '&:hover': {
+      //       backgroundColor: theme.palette.overlay.faint,
+      //     },
+      //   };
+      case 'light':
+        return {
+          backgroundColor: theme.palette.overlay.faint,
+          border: `1px solid ${theme.palette.overlay.subtle}`,
+          color: theme.palette.primary.contrastText,
+          '&:hover': {
+            backgroundColor: theme.palette.overlay.subtle,
+          },
+        };
+      default:
+        return {
+          backgroundColor: theme.palette.background.paper,
+          border: 'none',
+          color: theme.palette.primary.main,
+          fontWeight: theme.typography.button.fontWeight,
+          '&:hover': {
+            backgroundColor: theme.palette.background.default,
+          },
+        };
+    }
+  };
+
+  return {
+    padding: '8px 16px',
+    borderRadius: theme.shape.borderRadius.small,
+    cursor: 'pointer',
+    transition: theme.transitions.quick,
+    fontFamily: theme.typography.fontFamily,
+    fontSize: theme.typography.button.fontSize,
+    ...getStyles(),
+  };
+});
+
+const WelcomeText = styled('span')(({ theme }) => ({
+  color: theme.palette.overlay.nearWhite,
+  fontSize: theme.typography.body1.fontSize,
+  fontFamily: theme.typography.fontFamily,
+}));
+
+// -------------------- COMPONENT LOGIC --------------------
 function TopNavBar({ onLoginClick }) {
-    // Get authentication state and navigation function
-    const { user, logout } = useAuth();
-    const navigate = useNavigate();
+  const { user, logout } = useAuth();
+  const navigate = useNavigate();
 
-    /**
-     * Handle user logout
-     * Clears auth state and redirects to home
-     */
-    const handleLogout = async () => {
-        try {
-            // Call logout endpoint
-            await axiosInstance.post('/api/logout/');
-            // Clear auth context
-            logout();
-            // Navigate to home
-            navigate('/');
-        } catch (error) {
-            console.error('Logout error:', error);
-            // Still clear local state even if server call fails
-            logout();
-            navigate('/');
-        }
-    };
+  const handleLogout = async () => {
+    try {
+      await axiosInstance.post('/api/logout/');
+      logout();
+      navigate('/');
+    } catch (error) {
+      console.error('Logout error:', error);
+      logout();
+      navigate('/');
+    }
+  };
 
-    return (
-        <NavBar>
-            <NavLogo onClick={() => navigate('/')}>
-                <NavLogoImage src="/logo.png" alt="HCC Logo" />
-                <NavTitle>HCC Study Abroad</NavTitle>
-            </NavLogo>
+  return (
+    <NavBar>
+      <NavLogo onClick={() => navigate('/')}>
+        <NavLogoImage src="/logo.png" alt="HCC Logo" />
+        <NavTitle>HCC Study Abroad</NavTitle>
+      </NavLogo>
 
-            <NavControls>
-                {user ? (
-                    <>
-                        <WelcomeText>Welcome, {user.display_name}!</WelcomeText>
-                        <NavButton variant="light" onClick={() => navigate('/dashboard')}>
-                            Dashboard
-                        </NavButton>
-                        <NavButton variant="transparent" onClick={handleLogout}>
-                            Logout
-                        </NavButton>
-                    </>
-                ) : (
-                    <NavButton onClick={onLoginClick}>
-                        Login
-                    </NavButton>
-                )}
-            </NavControls>
-        </NavBar>
-    );
+      <NavControls>
+        {user ? (
+          <>
+            <WelcomeText>Welcome, {user.display_name}!</WelcomeText>
+            <NavButton variant="light" onClick={() => navigate('/dashboard')}>
+              Dashboard
+            </NavButton>
+            <NavButton variant="light" onClick={handleLogout}>
+              Logout
+            </NavButton>
+          </>
+        ) : (
+          <NavButton onClick={onLoginClick}>Login</NavButton>
+        )}
+      </NavControls>
+    </NavBar>
+  );
 }
 
 export default TopNavBar;
